% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeDeComSet-class.R
\name{as.MeDeComSet}
\alias{as.MeDeComSet}
\title{as.MeDeComSet}
\usage{
as.MeDeComSet(object, cg_subsets = 1, Ks = NULL)
}
\arguments{
\item{object}{An object of class \code{RefFreeCellMix} containing cell type deconvolution information, or a list of such objects.}

\item{cg_subsets}{The CpG subsets used in the analysis.}

\item{Ks}{The values of K used in the analysis. If NULL, K is determined by the size of the matrices.}
}
\value{
An object of type \code{MeDeComSet}
}
\description{
Function to convert object of type RefFreeCellMix to MeDeComSet
}
\details{
Since \code{RefFreeCellMix} only contains information on a single value for K, and does not contain any regularization
          (lambda), the corresponding parameters in the MeDeComSet are set to single numeric values. Furthermore, no information
          on goodness of fit (CVE, Fval) can be stored. If \code{cg_subsets} is not of length 1, an object containing multiple
          subsets is creared.
}
