% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\docType{data}
\name{AVAIL.REFS}
\alias{AVAIL.REFS}
\title{reference.R
-------------------------------------------------------------------------------------------------------------------------------------
This scripts contains functions for assocating MeDeCom with known reference cell type profiles
GLOBALS}
\format{
An object of class \code{character} of length 2.
}
\usage{
AVAIL.REFS
}
\description{
reference.R
-------------------------------------------------------------------------------------------------------------------------------------
This scripts contains functions for assocating MeDeCom with known reference cell type profiles
GLOBALS
}
\keyword{datasets}
