% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeDeCom-package.R
\docType{data}
\name{example.dataset}
\alias{example.dataset}
\title{Example dataset}
\usage{
data(example.dataset)
}
\description{
Contains a data set of methylation values as a \code{matrix} with 10,000 rows (CpGs) and 100 columns (samples), a reference of
contributions of 5 cell types in the samples, and the reference methylomes of the samples as a matrix with 10,000 rows and 5
columns.
 
\itemize{
          \item D, a \code{matrix} with 10,000 rows (CpGs) and 100 columns (samples) representing a potential input methylation
          matrix
          \item Aref, a reference of contributions of 5 cell types in the samples
          \item Tref, reference methylomes of the samples as a matrix with 10,000 rows and 5 columns
}
}
\author{
Michael Scherer
}
\keyword{datasets}
