% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeDeCom-package.R
\docType{data}
\name{example.MeDeComSet}
\alias{example.MeDeComSet}
\title{Example result object (MeDeComSet)}
\format{\code{MeDeComSet}, the result of a deconvolution experiment and input to any follow-up analysis or to retrieve latent 
methylation components (LMCs)}
\description{
An example result of applying \pkg{MeDeCom} for a deconvolution experiment.
}
\author{
Michael Scherer
}
\keyword{datasets}
