% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorizations.R
\name{factorize.alternate}
\alias{factorize.alternate}
\title{factorize.alternate}
\usage{
factorize.alternate(
  D,
  k,
  method = "MeDeCom.quadPen",
  t.method = "quadPen",
  Tfix = NULL,
  Tpartial = NULL,
  Tpartial.rows = NULL,
  Apartial = NULL,
  Apartial.cols = NULL,
  V = NULL,
  lambda = 0,
  init = "random",
  opt = 5,
  emp.dim = 500,
  emp.resample = TRUE,
  emp.vsf = 1,
  emp.borders = c(0, 1),
  qp.rangeT = c(0, 1),
  qp.Alower = NULL,
  qp.Aupper = NULL,
  itermax = 100,
  trace = FALSE,
  eps = 1e-08,
  ncores = 1,
  pheno = NULL,
  na.values = FALSE,
  seed = NULL,
  verbosity = 0L
)
}
\arguments{
\item{D}{m by n input matrix with mixture data}

\item{k}{number of latent components, \code{integer}}

\item{method}{optimization method used. Currently supported values are
\code{"MeDeCom.quadPen"} and \code{"MeDeCom.cppTAfact"}.}

\item{t.method}{method for updating the latent component matrix, one of 
\code{"integer", "empirical", "Hlasso"} or \code{"quadPen"}}

\item{Tfix}{an optional matrix of a priori known fixed components}

\item{V}{for \code{t.method} "integer" a small vector of possible values;
for \code{t.method} "empirical" a vector giving empirical distribution
of T values}

\item{init}{type of initialization, either "random" (default) or "fixed".}

\item{opt}{if \code{init} is "random" 
number of runs with independent initialization,
if \code{init} is "fixed"
starting values for T and A (see details)}

\item{emp.dim}{for \code{t.method} "empirical", 
number of randomly drawn samples for T row selection}

\item{emp.resample}{for \code{t.method} "empirical",
a flag indicating whether resampling should
be done at each iteration}

\item{itermax}{maximal number of iterations}

\item{trace}{a flag indicating whether to return the 
factorization results for each iteration}

\item{eps}{threshold for objective value change}

\item{ncores}{number of CPU cores used for parallelization}

\item{pheno}{a list with phenotypic information}

\item{verbosity}{flag specifying whether to show diagnostic
statements during the execution}
}
\value{
a \code{list} with the following elements:
			\describe{
	 				\item{\code{T}}{matrix of latent components}
  					\item{\code{A}}{matrix of mixture proportions}
						\item{\code{Fval}}{the final value of the objective function}
						\item{\code{Conv}}{sequence of objective function values
								attained after each iteration} 
						\item{\code{rmse}}{RMSE of the factorization}
			}
}
\description{
Matrix factorization algorithms based on the alternating optimization scheme
}
\details{
In case \code{init} is "fixed" the starting values
						for the m by k matrix of latent components 
						and for the k by n matrix of mixing proportions 
						should be specified as T and A elements of a list
						supplied as \code{opt}
}
\author{
Martin Slawski

R port by Pavlo Lutsik
}
