% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorizations.R
\name{factorize.regr}
\alias{factorize.regr}
\title{factorize.regr}
\usage{
factorize.regr(D, Tt, A0 = NULL, precision = 1e-08)
}
\arguments{
\item{D}{m by n matrix with mixture data}

\item{Tt}{either a m by k matrix of k true latent components
or a list of n such matrices, one per each column of D}

\item{A0}{initialization for the mixture proportions matrix}

\item{precision}{numerical tolerance of the optimization algorithm}
}
\value{
a \code{list} with elements:
					\describe{
						\item{\code{A}}{matrix of mixing proportions}
						\item{\code{T}}{Tt used}
						\item{\code{rmse}}{RMSE of the regression model}
					}
}
\description{
Get mixing proportions from the target data matrix and a matrix of latent factors
}
