% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeDeComSet-class.R
\docType{methods}
\name{getProportions,MeDeComSet-method}
\alias{getProportions,MeDeComSet-method}
\alias{getProportions}
\title{getProportions-methods}
\usage{
\S4method{getProportions}{MeDeComSet}(object, K = object@parameters$Ks[1],
  lambda = object@parameters$lambdas[1], cg_subset = 1, statistic = "cve")
}
\arguments{
\item{object}{object returned by \link{runMeDeCom}}

\item{K}{number of LMCs}

\item{lambda}{regularlization parameter}

\item{cg_subset}{used CpG subset, defaults to the full data set}

\item{statistic}{statistic to be used in returning}
}
\description{
Return a matrix of LMCs
}
\examples{
\donttest{
data(example.data)
getProportions(example_MeDeComSet, K=2, lambda=0.001)
}
}
