% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{greedymatch}
\alias{greedymatch}
\title{greedymatch}
\usage{
greedymatch(Tstar, That, Ahat)
}
\description{
Matching of latent components
}
\details{
suppose Tstar contains K* topics (columns)
1. select the K* most popular topics of That
2. use a greedy method to match them to those of Tstar (w.r.t L2 norm)
   and correspondingly permute rows of Ahat
}
\author{
Martin Slawski

R port by Pavlo Lutsik
}
