% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMC_interpretation.R
\name{lmc.annotation.enrichment}
\alias{lmc.annotation.enrichment}
\title{lmc.annotation.enrichment}
\usage{
lmc.annotation.enrichment(
  medecom.result,
  annotation.filter = NULL,
  anno.data,
  K = NULL,
  lambda = NULL,
  cg_subset = NULL,
  diff.threshold = 0.5,
  reference.computation = "median",
  comp.lmcs = NULL,
  type = "hypo",
  assembly = "hg19"
)
}
\arguments{
\item{medecom.result}{An object of type \code{\link{MeDeComSet-class}} or the location of an .RData file, where
such an object is stored.}

\item{annotation.filter}{A numeric vector specifying the sites that have been removed from \code{rnb.set} in a
preprocessing step (e.g. coverage filtering) or a path to an .RData file.}

\item{anno.data}{The original \code{\link[RnBeads]{RnBSet-class}} object containing methylation, sample meta and annotation
information or a path to a directory stored by \code{\link[RnBeads]{save.rnb.set}} or a data.frame containing
CpG annotations (ann_C)}

\item{K}{The number of LMCs specified for the MeDeCom run.}

\item{lambda}{The lambda parameter selected.}

\item{cg_subset}{The index of the selection strategy employed (e.g. most variable CpGs).}

\item{diff.threshold}{The difference cutoff between median methylation in the remaining LMCs and the LMC of interest
used to call a CpG differentially methylated. The higher this value, the more conservative the
selection.}

\item{reference.computation}{Metric used to set the reference on the remaining LMCs to determine hyper- and hypomethylated sites.
Can be either \code{"median"} (default), \code{"mean"}, or \code{"lmcs"} (\code{comp.lmcs} argument needs to be provided).}

\item{comp.lmcs}{Numeric vector containing two numbers representing the LMCs that should be compared to one another.}

\item{type}{Which direction is to be tested for enrichment. Can be one of "hypo", "hyper", or "differential"}

\item{assembly}{The assembly used. Needs to be one of "hg19", "hg38" or "mm10". Does not need to be specified, if rnb.set is a
\code{\link{RnBSet-class}}}
}
\value{
A data frame with four columns: \describe{
 \item{LMC}{The LMC analyzed}
 \item{annotation}{The annotation used. Can either be \code{chrXY} for enrichments on different chromosmes or different functional categories in the Ensembl regulatory build}
 \item{p.value}{The p-value computing using Fisher's exact test for enrichment}
 \item{OR}{The odds ratio for enrichment}
}
}
\description{
This function performs enrichment analysis for various genomic locations including chromosomes,
and different function categories defined by the Ensembl regulatory build
}
\author{
Michael Scherer
}
