% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMC_interpretation.R
\name{lmc.go.enrichment}
\alias{lmc.go.enrichment}
\title{lmc.go.enrichment}
\usage{
lmc.go.enrichment(medecom.result, annotation.filter = NULL, rnb.set, K,
  lambda, cg_subset = NULL, diff.threshold = 0.5, region.type = "genes",
  temp.dir = tempdir(), type = "hypo")
}
\arguments{
\item{medecom.result}{An object of type \code{\link{MeDeComSet-class}} or the location of an .RData file, where
such an object is stored.}

\item{annotation.filter}{A numeric vector specifying the sites that have been removed from \code{rnb.set} in a
preprocessing step (e.g. coverage filtering) or a path to an .RData file.}

\item{rnb.set}{The original \code{\link[RnBeads]{RnBSet-class}} object containing methylation, sample meta and annotation
information or a path to a directory stored by \code{\link[RnBeads]{save.rnb.set}}.}

\item{K}{The number of LMCs specified for the MeDeCom run.}

\item{lambda}{The lambda parameter selected.}

\item{cg_subset}{The index of the selection strategy employed (e.g. most variable CpGs).}

\item{diff.threshold}{The difference cutoff between median methylation in the remaining LMCs and the LMC of interest
used to call a CpG differentially methylated. The higher this value, the more conservative the
selection.}

\item{region.type}{Region type used to annotate CpGs to potentially regulatory regions (see \url{https://rnbeads.org/regions.html})
for a list of available region types. Here, only "genes" "promoters" and their gencode versions
are available.}

\item{temp.dir}{Path to a directory used to store temporary files.}

\item{type}{Which direction is to be tested for enrichment. Can be one of "hypo", "hyper", or "differential"}
}
\value{
A list with K elements. One element is the enrichment result of the corresponding LMC-specific hypomethylated CpG sites.
}
\description{
This routine computes GO enrichment results for LMC-specifically hypo- or hypermethylated sites.
}
\details{
This function employs GO enrichment analysis with the GOstats package on the CpG sites that are LMC-specifically
                 hypomethylated, after annotating the sites to the closest promotor/gene defined by \code{region.type}.
                 The sites are selected by computing the median methylation value of the other LMCs and then selecting
                 those sites that are more than \code{diff.threshold} away from the median in the LMC of interest.
                 This is done for all LMCs from 1 to K.
}
\author{
Michael Scherer
}
