% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMC_interpretation.R
\name{lmc.go.plots.tables}
\alias{lmc.go.plots.tables}
\title{lmc.go.plots.tables}
\usage{
lmc.go.plots.tables(medecom.result, anno.data, ...)
}
\arguments{
\item{medecom.result}{An object of type \code{\link{MeDeComSet-class}} or the location of an .RData file, where
such an object is stored.}

\item{anno.data}{The original \code{\link[RnBeads]{RnBSet-class}} object containing methylation, sample meta and annotation
information, a path to a directory stored by \code{\link[RnBeads]{save.rnb.set}} or a data.frame containing
CpG annotations (ann_C)}

\item{...}{Further arguments passed to \code{lmc.lola.enrichment}}
}
\value{
A list with two elements, one of them containing the plots for each LMC and the other for the corresponding GO
        enrichment tables
}
\description{
This functions calls \link{lmc.go.enrichment} and returns plots representing those results, as well as the tables with GO
enrichment results.
}
\seealso{
lmc.go.enrichment
}
\author{
Michael Scherer
}
