% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMC_interpretation.R
\name{lmc.lola.enrichment}
\alias{lmc.lola.enrichment}
\title{lmc.lola.enrichment}
\usage{
lmc.lola.enrichment(medecom.result, annotation.filter = NULL, anno.data,
  K = NULL, lambda = NULL, cg_subset = NULL, diff.threshold = 0.5,
  reference.computation = "median", comp.lmcs = NULL,
  region.type = "ensembleRegBuildBPall", temp.dir = tempdir(),
  type = "hypo", assembly = "hg19", lola.db = NULL)
}
\arguments{
\item{medecom.result}{An object of type \code{\link{MeDeComSet-class}} or the location of an .RData file, where
such an object is stored.}

\item{annotation.filter}{A numeric vector specifying the sites that have been removed from \code{rnb.set} in a
preprocessing step (e.g. coverage filtering) or a path to an .RData file.}

\item{anno.data}{The original \code{\link[RnBeads]{RnBSet-class}} object containing methylation, sample meta and annotation
information, a path to a directory stored by \code{\link[RnBeads]{save.rnb.set}} or a data.frame containing
CpG annotations (ann_C)}

\item{K}{The number of LMCs specified for the MeDeCom run.}

\item{lambda}{The lambda parameter selected.}

\item{cg_subset}{The index of the selection strategy employed (e.g. most variable CpGs).}

\item{diff.threshold}{The difference cutoff between median methylation in the remaining LMCs and the LMC of interest
used to call a CpG differentially methylated. The higher this value, the more conservative the
selection.}

\item{reference.computation}{Metric used to set the reference on the remaining LMCs to determine hyper- and hypomethylated sites.
Can be either \code{"median"} (default), \code{"mean"}, or \code{"lmcs"} (\code{comp.lmcs} argument needs to be provided).}

\item{comp.lmcs}{Numeric vector containing two numbers representing the LMCs that should be compared to one another.}

\item{region.type}{Region type used to annotate CpGs to potentially regulatory regions (see \url{https://rnbeads.org/regions.html})
for a list of available region types.}

\item{temp.dir}{Path to a directory used to store temporary files.}

\item{type}{Which direction is to be tested for enrichment. Can be one of "hypo", "hyper", or "differential"}

\item{assembly}{The assembly used. Needs to be one of "hg19", "hg38" or "mm10". Does not need to be specified, if rnb.set is a
\code{\link{RnBSet-class}}}

\item{lola.db}{A loaded LOLA database as loaded with LOLA::loadRegionDB. If this value is NULL, the database is loaded
automatically and stored in the temporary directory.}
}
\value{
A list with K elements. One element is the enrichment result of the corresponding LMC-specific hypomethylated CpG sites.
}
\description{
This routine computes LOLA enrichment results for LMC-specifically hypo- or hypermethylated sites.
}
\details{
This function employs LOLA on the CpG sites that are LMC-specifically hypomethylated, after annotating 
                 the sites to the closest region defined by \code{region.type}. The sites are selected by computing
                 the median methylation value of the other LMCs and then selecting those sites that are more than 
                 \code{diff.threshold} away from the median in the LMC of interest. This is done for all LMCs from
                 1 to K.
}
\seealso{
lmc.lola.plot.tables
}
\author{
Michael Scherer
}
