% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMC_interpretation.R
\name{lmc.lola.plots.tables}
\alias{lmc.lola.plots.tables}
\title{lmc.lola.plots.tables}
\usage{
lmc.lola.plots.tables(medecom.result, annotation.filter = NULL,
  anno.data, K = NULL, lambda = NULL, cg_subset = NULL,
  diff.threshold = 0.5, region.type = "ensembleRegBuildBPall",
  temp.dir = tempdir(), type = "hypo", assembly = "hg19",
  lola.db = NULL)
}
\arguments{
\item{medecom.result}{An object of type \code{\link{MeDeComSet-class}} or the location of an .RData file, where
such an object is stored.}

\item{annotation.filter}{A numeric vector specifying the sites that have been removed from \code{rnb.set} in a
preprocessing step (e.g. coverage filtering) or a path to an .RData file.}

\item{anno.data}{The original \code{\link[RnBeads]{RnBSet-class}} object containing methylation, sample meta and annotation
information, a path to a directory stored by \code{\link[RnBeads]{save.rnb.set}} or a data.frame containing
CpG annotations (ann_C)}

\item{K}{The number of LMCs specified for the MeDeCom run.}

\item{lambda}{The lambda parameter selected.}

\item{cg_subset}{The index of the selection strategy employed (e.g. most variable CpGs).}

\item{diff.threshold}{The difference cutoff between median methylation in the remaining LMCs and the LMC of interest
used to call a CpG differentially methylated. The higher this value, the more conservative the
selection.}

\item{region.type}{Region type used to annotate CpGs to potentially regulatory regions (see \url{https://rnbeads.org/regions.html})
for a list of available region types.}

\item{temp.dir}{Path to a directory used to store temporary files.}

\item{type}{Which direction is to be tested for enrichment. Can be one of "hypo", "hyper", or "differential"}

\item{assembly}{The assembly used. Needs to be one of "hg19", "hg38" or "mm10". Does not need to be specified, if rnb.set is a
\code{\link{RnBSet-class}}}

\item{lola.db}{A loaded LOLA database as loaded with LOLA::loadRegionDB. If this value is NULL, the database is loaded
automatically and stored in the temporary directory.}
}
\value{
A list with two elements, one of them containing the plots for each LMC and the other for the corresponding LOLA
        enrichment tables
}
\description{
This functions calls \link{lmc.lola.enrichment} and returns plots representing those results, as well as the tables with LOLA
enrichment results.
}
\seealso{
lmc.lola.enrichment
}
\author{
Michael Scherer
}
