% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMC_interpretation.R
\name{load.lola.for.medecom}
\alias{load.lola.for.medecom}
\title{load.lola.for.medecom}
\usage{
load.lola.for.medecom(dir.path = tempdir(), assembly = "hg19")
}
\arguments{
\item{dir.path}{A path to a directory, where the LOLA database is to be downloaded. Defaults to the temporary directory.}

\item{assembly}{The assembly to be used.}
}
\value{
The loaded LOLA database
}
\description{
This functions downloads and loads the LOLA database in the specified directory. Should only be called once per session to save time.
}
\author{
Michael Scherer
}
