% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{locus_plot}
\alias{locus_plot}
\title{locus_plot}
\usage{
locus_plot(That, ann, cgs, locus.chr, locus.start, locus.end,
  locus.name = sprintf("\%s:\%d-\%d", locus.chr, locus.start, locus.end),
  locus.forward = FALSE, flank.start = 1000, flank.end = 1000,
  comp.cols = rainbow(ncol(That)), legend.pos = "topleft",
  Tstar = NULL, D = NULL, plot.genes = FALSE, ann.genes = NULL)
}
\arguments{
\item{That}{a matrix of LMCs}

\item{ann}{CpG annotation}

\item{cgs}{indices of CpGs data for which are present in That with respect to ann}

\item{locus.chr}{chromosome}

\item{locus.start}{start coordinate}

\item{locus.end}{end coordinate}

\item{flank.start}{number of basepairs to extend the locus upstream}

\item{flank.end}{number of basepairs to extend the locus downstream}

\item{comp.cols}{color code for LMCs}

\item{legend.pos}{location of the legend, in accordance with \link{legend}}

\item{Tstar}{matrix of reference profiles}

\item{D}{matrix of input methylation data used to produce That}

\item{plot.genes}{if \cs{TRUE} a track with gene locations will be plotted}

\item{ann.genes}{gene annotation necessary for the gene plotting

 @author Pavlo Lutsik, with modifications by Michael Scherer
 @export}
}
\description{
Plot results for a selected locus
}
