% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{matchLMCs}
\alias{matchLMCs}
\title{match.components}
\usage{
matchLMCs(That, Tref, method = "corrmatch", check = TRUE, Ahat = NULL)
}
\arguments{
\item{That}{recovered components from factorization}

\item{Tref}{reference components}

\item{method}{one of "corrmatch" and "greedymatch"}

\item{check}{for method "corrmatch": a flag specifying whether to check uniquenes of the match}

\item{Ahat}{for method "greedymatch": recovered mixing proportions}
}
\value{
a vector of indices. The length and the order of the vector corresponds to the columns of \code{TT}
					and the indices specify the columns of Tref
}
\description{
Matching the recovered components to the reference
}
\details{
Wrapper function for component matching methods
}
