% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeDeCom-package.R
\docType{package}
\name{MeDeCom-package}
\alias{MeDeCom-package}
\title{MeDeCom: Methylome DeComposition using regularized constrained matrix factorization}
\description{
MeDeCom is an R-package that discovers and quantifies latent components
in the DNA methylomes of heterogeneous samples
}
\references{
TBA
}

