% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeDeComSet-class.R
\docType{class}
\name{MeDeComSet-class}
\alias{MeDeComSet-class}
\alias{MeDeComSet}
\alias{initialize,MeDeComSet-method}
\title{MeDeComSet Class}
\usage{
MeDeComSet(parameters, outputs, dataset_info = list())
}
\arguments{
\item{parameters}{\code{list} of MeDeCom parameters with elements \code{K}, integer vector of k values, \code{lambdas}, numeric vector of lambda values.}

\item{outputs}{\code{list} of MeDeCom resutls with one element per each used CpG subset.}

\item{dataset_info}{\code{list} with information about the input data set.}
}
\value{
an object of class MeDeComSet
}
\description{
Stores the result of a methylation deconvolution experiment

Wrapper function MeDeComSet
}
\section{Slots}{

\describe{
  \item{\code{dataset_info}}{\code{list} with information about the input data set.}
  \item{\code{parameters}}{\code{list} containing parameters of the deconvolution experiment.}
  \item{\code{outputs}}{\code{list} of deconvolution products for each combination of MeDeCom parameters.}
}
}

\section{Methods and functions}{

\describe{
  \item{\code{\link[=getStatistics,MeDeComSet-method]{getStatistics}}}{Returns the value of one goodness-of-fit statistics for a given parameter combination.}
  \item{\code{\link[=getLMCs,MeDeComSet-method]{getLMCs}}}{Returns a matrix of LMCs for a given parameter combination.}
  \item{\code{\link[=getProportions,MeDeComSet-method]{getProportions}}}{Returns a matrix of mixing proportions for a given parameter combination.}
  \item{\code{\link{plotParameters}}}{Create a parameter selection plot.}
  \item{\code{\link{plotLMCs}}}{Visualize the LMCs.}
  \item{\code{\link{plotProportions}}}{Visualize the mixing proportions.}
}
}

\author{
Pavlo Lutsik
}
