% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{plotLMC.reference}
\alias{plotLMC.reference}
\title{plotLMC.reference}
\usage{
plotLMC.reference(medecom.set, ref.meth, ann.md,
  cg_subset = medecom.set@parameters$cg_subsets[1],
  K = medecom.set@parameters$Ks[1],
  lambda = medecom.set@parameters$lambdas[1], ann.ref = NULL,
  type = "dendrogram", chrom.col = "Chromosome", start.col = "Start",
  end.col = "End")
}
\arguments{
\item{medecom.set}{An object of type MeDeComSet}

\item{ref.meth}{Reference methylome in the form of a matrix of a RnBSet}

\item{ann.md}{Genomic annotation for the CpG sites present in medecom.set}

\item{cg_subset}{The cg_subset of interest}

\item{K}{K value}

\item{lambda}{lambda value}

\item{ann.ref}{Genomic annotation for the CpG sites in the referenc methylome. Can be omitted, if ref.meth is an RnBSet object}

\item{type}{Plot type, see \code{\link{plotLMC}}}

\item{chrom.col}{The chromosome column name in ann.md and ann.ref}

\item{start.col}{The start column name in ann.md and ann.ref}

\item{end.col}{The end column name in ann.md and ann.ref}
}
\description{
This routine uses a reference methylome, together with an annotation, and calls plotLMCs for the matching postions
}
\author{
Michael Scherer
}
