% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotLMCs}
\alias{plotLMCs}
\title{plotLMCs}
\usage{
plotLMCs(MeDeComSet, type, K = NA, lambda = NA, cg_subset = 1,
  lmc = NA, Tref = NULL, distance = "correlation", center = FALSE,
  n.most.var = NA, D = NULL, sample.characteristic = NULL,
  scatter.matching = FALSE, scatter.smooth = TRUE,
  scatter.cg.feature = NULL, min.similarity = 0)
}
\arguments{
\item{MeDeComSet}{an object with MeDeCom results}

\item{type}{plot type, a \code{character} of length 1 (see Details)}

\item{K}{value of parameter k to use}

\item{lambda}{value of parameter lambda to use}

\item{cg_subset}{which CpG subset to use}

\item{lmc}{which LMC to use for visualization}

\item{Tref}{a matrix with reference methylomes}

\item{distance}{distance measure to use}

\item{center}{if \code{TRUE} the LMC and reference methylome matrices will be row-centered}

\item{n.most.var}{is not \code{NA} a respective number of CpGs with the highest standard deviation will be plotted}

\item{D}{input data matrix used to derive the LMCs}

\item{min.similarity}{minimal similarity between LMCs and (if available) reference profiles, used to select edges in \code{"similarity graph"}.
Has only an influence if \code{type}=\code{"similarity graph"}.}
}
\description{
A wrapper for various plotting methods for the visualization of LMCs
}
\details{
Available plot types include:
\describe{
\item{\bold{\code{dendrogram}}}{
       Dendrogram visualizing a joint hierarchical clustering of LMCs and, if available, the reference methylomes.}
\item{\bold{\code{heatmap}}}{
       Heatmap visualizing a distance between LMCs and the reference methylomes.}
\item{\bold{\code{mds}}}{
       Joint multidimensional scaling of LMCs and the reference methylomes.}
\item{\bold{\code{scatterplot}}}{
       Multi-panel scatterplot of LMCs and reference methylomes.}
\item{\bold{\code{extremality}}}{
       Barplot visualizing the value of the regularizer term for each LMC.}
\item{\bold{\code{distance to center}}}{
       Barplot visualizing a distance to the data center for each LMC. Input data matrix used to derive the LMCs should be 
		  supplied as argument \code{D}.}
}
}
