% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotParameters}
\alias{plotParameters}
\title{plotParameters}
\usage{
plotParameters(MeDeComSet, cg_subset = 1, Ks = integer(),
  lambdas = integer(), statistic = "cve", minLambda = 0,
  maxLambda = Inf, lambdaScale = "native", ...)
}
\arguments{
\item{MeDeComSet}{MeDeCom object}

\item{cg_subset}{integer index of the CpG subset (if no subsets were used)}

\item{Ks}{values of parameter k to use (by default all k values available in the \code{MeDeComSet} are used)}

\item{lambdas}{values of parameter lambda to use (by default all lambda values available in the \code{MeDeComSet} are used)}

\item{statistic}{if multiple k values are supplied, the statistic which is plotted (defaults to cross-validation error)}

\item{minLambda}{minimal lambda value}

\item{maxLambda}{maximal lambda value}
}
\description{
Parameter selection plots
}

