% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotProportions}
\alias{plotProportions}
\title{plotProportions}
\usage{
plotProportions(MeDeComSet, type, K, lambda, cg_subset = 1, lmc = NA,
  Aref = NULL, ref.profile = NA, assignment.method = "pearson",
  sample.characteristic = NULL, heatmap.clusterCols = FALSE,
  heatmap.clusterRows = FALSE)
}
\arguments{
\item{MeDeComSet}{an object with MeDeCom results}

\item{type}{plot type, a \code{character} of length 1 (see Details)}

\item{K}{value of parameter k to use}

\item{lambda}{value of parameter lambda to use}

\item{cg_subset}{which CpG subset to use}

\item{lmc}{which LMC to use for visualization}

\item{Aref}{a matrix with reference methylomes}
}
\description{
A wrapper for various plotting methods for the visualization of mixing proportions
}
\details{
Available plot types include:
\describe{
\item{\bold{\code{heatmap}}}{
       Lineplot of proportions recovered by MeDeCom and reference proportions.}
\item{\bold{\code{barplot}}}{
       Stacked barplot of proportions recovered by MeDeCom.}
\item{\bold{\code{lineplot}}}{
       Lineplot of proportions recovered by MeDeCom and, if available, reference proportions.}
\item{\bold{\code{scatterplot}}}{
       Lineplot of proportions recovered by MeDeCom and reference proportions.}
}
}
