% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{run.refbased}
\alias{run.refbased}
\title{FUNCTIONS
run.refbased}
\usage{
run.refbased(
  rnb.set,
  Ks,
  lambdas,
  cg_subsets = NULL,
  opt.method = "MeDeCom.cppTAfact",
  temp.dir = NULL,
  ref.base = "reinius",
  most.var = NULL,
  NCORES = 1,
  cluster.settings = NULL,
  ref.set = NULL,
  id.col = NULL,
  save.restricted.sites = FALSE
)
}
\arguments{
\item{rnb.set}{An Object of type \code{RnBSet} from the \pkg{RnBeads} package containing methylation information on which
the deconvolution is to be performed}

\item{Ks}{Numeric vector containing the number of components to be computed by \pkg{MeDeCom}}

\item{lambdas}{Numeric vector specifying the regularzation parameters to be explored}

\item{cg_subsets}{List of numeric vectors specifying the rows that are to be used in the analysis.}

\item{opt.method}{Optimization method to be employed. For further information see \code{\link{runMeDeCom}}}

\item{temp.dir}{Optional temporary directory to store intermediate results}

\item{ref.base}{Reference profile data base to be used. Supported are 
\itemize{
  \item \code{"reinius"} A blood cell type reference methylome data set from Reinius et.al. (Reference to be added)
}
\itemize{
  \item \code{"local"} The reference data set is provided by the user. If this option is selected, \code{ref.set} 
                     must not be empty.
}}

\item{most.var}{Number specifying the number of most variable to be selected from \code{rnb.set}}

\item{NCORES}{Number of cores to be used for analysis.}

\item{cluster.settings}{Setting for the environment of a high performance compute cluster. Passed to \code{\link{runMeDeCom}}}

\item{ref.set}{A \code{RnBSet} object containing a reference data set to be used besides \code{rnb.set}. Is only compatible with
\code{ref.base="local"}.}

\item{id.col}{The name of the column in the sample annotation sheet of \code{ref.set} containing the reference cell type. Is only
compatible with \code{ref.base="local"}.}

\item{save.restricted.sites}{Flag indicating if \code{rnb.set} restricted to the \code{most.var} sites is to be saved in \code{temp.dir}
for potential downstream analysis.}
}
\value{
A list object containing two elements \itemize{
           \item \code{"MeDeComSet"} Results of applying MeDeCom with the setting above
           \item \code{"RefMeth"} A matrix containing reference profiles from the specified data set. The number of rows in 
                       this matrix has been reduced according to the most variable sites in \code{rnb.set}.
        }
}
\description{
Function to link MeDeCom's output to reference methylation profiles. The function returns both the MeDeCom result and a methyalation
matrix with reference profiles from the source specified.
}
\details{
This function applied MeDeCom to the specified data set and only support \code{RnBSet} objects as inputs. The function
 internally manipulated the object by selecting the most variable sites according to \code{most.var}. This leads to a decrease in
 the number of rows in the reference profiles to this number. 
 
 Please note that an active internet connection is required, since this routine downloads data through the world wide web.
}
\author{
Michael Scherer
}
