% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribution_interpretation.R
\name{run.trait.association.single}
\alias{run.trait.association.single}
\title{run.trait.association.single}
\usage{
run.trait.association.single(medecom.set, rnb.set, cg_subset = NULL,
  K = NULL, lambda = NULL, test.fun = t.test)
}
\arguments{
\item{medecom.set}{An object of type \code{\link{MeDeComSet}} as the result of \code{\link{runMeDeCom}} containing LMCs and their
proportions in the samples. The Set can contain multiple runs for different values of K and lambda.}

\item{rnb.set}{An object of type \code{\link[RnBeads]{RnBSet-class}} containing methylation data and metadata for the same samples for which 
\code{medecom.set} was computed or a data.frame of sample annotations (ann_S)}

\item{cg_subset}{The cg_subset of interest}

\item{K}{The selected value for number of LMCs (K)}

\item{lambda}{The selected value of the regularizer (lambda)}

\item{test.fun}{Test statistic used to compute p-values of differences between LMC contributions in pairwise sample comparisons.
Defaults to \code{t.test}.}

\item{plot.path}{Path to store the p-value heatmaps.}

\item{figure.format}{Character describing the format in which plots should be stored on disk. Either \code{"pdf"} or \code{"png"}.}
}
\description{
Computes test statistics for all possible group assignments of samples defined in \code{medecom.set} and \code{rnb.set} and stores
heatmaps of p-values on the given location only for a given CG Subset, K and lambda.
}
\details{
This function creates a new folder names \code{pdfs} at the location given by \code{plot.path} and stores a heatmap for
          all possible Ks and lambdas defined in \code{medecom.set}. The p-values are produced by comparing the LMC contributions
          in all sample comparisons defined by \code{\link[RnBeads]{rnb.sample.groups}} on \code{rnb.set}. The employed test statistic for
          pariwise comparison can be specified by \code{test.fun}, for groups defining more than one group \code{\link{kruskal.test}}
          is employed. P-values lower than 0.01 are added to the heatmap.
}
\author{
Michael Scherer
}
