% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend.R
\name{runMeDeCom}
\alias{runMeDeCom}
\title{runMeDeCom}
\usage{
runMeDeCom(data, Ks, lambdas, opt.method = "MeDeCom.cppTAfact",
  cg_subsets = NULL, sample_subset = NULL, startT = NULL, startA = NULL,
  trueT = NULL, trueA = NULL, fixed_T_cols = NULL, NINIT = 100,
  ITERMAX = 1000, NFOLDS = 10, N_COMP_LAMBDA = 4, NCORES = 1,
  random.seed = NULL, num.tol = 1e-08, analysis.name = NULL,
  use.ff = FALSE, cluster.settings = NULL, temp.dir = NULL,
  cleanup = TRUE, verbosity = 1L, time.stamps = FALSE)
}
\arguments{
\item{data}{DNA methylation dataset as a \code{numeric} matrix (methylation sites vs samples) or an ojbect of class \code{RnBeadSet}}

\item{Ks}{values of parameter \deqn{k} to be tested, vector of type \code{integer}}

\item{lambdas}{values of parameter \deqn{\lambda} to be tested, vector of type \code{numeric}}

\item{opt.method}{optimization method used. Currently supported values are \code{"MeDeCom.quadPen"} and \code{"MeDeCom.cppTAfact"}.}

\item{cg_subsets}{a \code{list} of \code{integer} vectors specifying row indices to include into the analysis}

\item{sample_subset}{samples to include into the analysis}

\item{startT}{a \code{list} of length equal to \code{length(Ks)} or a \code{matrix} with \code{max(Ks)} columns}

\item{startA}{a \code{list} of length equal to \code{length(Ks)} or a \code{matrix} with \code{max(Ks)} rows}

\item{trueT}{a numeric matrix with as many rows as there are methylation sites in \code{data}}

\item{trueA}{a numeric matrix with as many columns as there are methylation sites in \code{data}}

\item{fixed_T_cols}{columsn of T which are known (to be implemented)}

\item{NINIT}{number of random initializations}

\item{ITERMAX}{maximal number of iterations of the alternating optimization scheme}

\item{NFOLDS}{number of cross-validation folds}

\item{N_COMP_LAMBDA}{the number of solutions to compare in the "smoothing" step}

\item{NCORES}{number of cores to be used in the parallelized steps (at best a divisor of NINIT)}

\item{random.seed}{seed for random number generation}

\item{num.tol}{some small parameter}

\item{analysis.name}{a deliberate name of the analysis as a \code{character} singleton}

\item{use.ff}{use \code{ff} package functionality for memory optimization}

\item{cluster.settings}{a list with parameters for an HPC cluster}

\item{temp.dir}{a temporary directory for the cluster-based analysis available on all nodes}

\item{cleanup}{if \code{TRUE} the temporary directory will be removed on completion}

\item{verbosity}{verbosity level, \code{0} for quiet execution}

\item{time.stamps}{add timestamps to the diagnostic output}
}
\value{
MeDeComSet object
}
\description{
Perform a MeDeCom experiment
}
\author{
Pavlo Lutsik
}
