% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{cluster.refbased}
\alias{cluster.refbased}
\title{cluster.refbased}
\usage{
cluster.refbased(ref.run, K, lambda, cg_subset = 1,
  plot.type = "dendrogram", ...)
}
\arguments{
\item{ref.run}{A result of \code{\link{run.refbased}} with the results of \code{\link{runMeDeCom}} and the reference data matrix}

\item{K}{Selected number of components}

\item{lambda}{Selected regularization parameter}

\item{cg_subset}{Numeric vector representing the CpG sites selected for analysis from the orignial MeDeComSet}

\item{plot.type}{Type of plot to be created, is passed to \code{\link{plotLMCs}}}

\item{...}{Further arguments passed to plotLMCs}
}
\value{
A plot object displaying the clustering
}
\description{
This routine performs general hierachical clustering as in the \code{\link{plotLMCs}} and adds reference methylome profiles to this
clustering.
}
\author{
Michael Scherer
}
